merge in fusion experiments

See merge request smart/watch!83

+++ watch/tasks/fusion/predict.py
@@ -7,9 +7,9 @@ import kwcoco
 import kwimage
 from torch import nn
 
-from . import datasets
-from . import methods
-from . import utils
+from watch.tasks.fusion import datasets
+from watch.tasks.fusion import methods
+from watch.tasks.fusion import utils
 
 fname_template = "{location}/{bands}-{frame_no}.tif"
 
@@ -71,7 +71,7 @@ def main(args):
 
     # fill canvases
     for example, meta in zip(tqdm.tqdm(test_dataloader), test_dataset.sample_grid):
-        images, labels = example["images"], example["labels"]
+        images, labels = example["images"].float(), example["labels"]
         changes = (labels[0, 1:] != labels[0, :-1]).detach().cpu().numpy()
         T, H, W = changes.shape
